apps/cic-eth/tools_requirements.txt
crypto-dev-signer>=0.4.14b7,<=0.4.14
chainqueue>=0.0.3a1,<0.1.0
confini>=0.3.6rc4,<0.5.0
cic-eth-registry>=0.5.8a1,<0.6.0
redis==3.5.3
hexathon~=0.0.1a7
pycryptodome==3.10.1
pyxdg==0.27
